refactor Keycloak-specific auth to work with any OpenID server
This allows us to use Cognito with a config change rather than
a big yucky refactor
