+++ server/cdk/lib/cdk-stack.ts
@@ -67,7 +67,8 @@ export class CdkStack extends cdk.Stack {
       DB_HOST: db.dbInstanceEndpointAddress,
       DB_NAME: "TODO",
       DB_USER: "postgres",
-      AUTH_HOST: config.authProviderHostname,
+      // https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-userpools-server-contract-reference.html
+      OIDC_ENDPOINT: `https://cognito-idp.${this.region}.amazonaws.com/${userPool.userPoolId}`,
       EXPO_ACCESS_TOKEN: config.expoAccessToken ?? "",
     };
     const userData = readFileSync(path.join(cwd(), "assets", "userdata.sh"))
