+++ server/docker-compose.yaml
@@ -7,7 +7,7 @@ services:
       - DB_NAME
       - DB_USER
       - DB_PASSWORD
-      - KEYCLOAK_HOST=keycloak:8080
+      - OIDC_ENDPOINT=http://keycloak:8080
       - EXPO_ACCESS_TOKEN
     ports:
       - "3000"
